requirements.txt
asgiref==3.7.2
attrs==23.2.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==4.1.13
django-cors-headers==3.10.1
django-filter==23.5
djangorestframework==3.14.0
djlint==1.34.1
djongo==1.3.6
dnspython==2.6.0
drf-spectacular==0.27.1
EditorConfig==0.12.4
gunicorn==20.0.4
html-tag-names==0.1.2
html-void-elements==0.1.0
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
mysqlclient==2.2.4
pathspec==0.12.1
pillow==10.2.0
pymongo==3.12.1
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2019.3
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
rpds-py==0.18.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.2.4
tqdm==4.66.2
tzdata==2024.1
uritemplate==4.1.1
wheel==0.41.2
